+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Removed "codegen" module that was an experimental annotation based code generator
 - Added [JSpecify](https://jspecify.dev/) and [NullAway](https://github.com/uber/nullaway)
 - Use encryption functionality from [event-store-commons](https://github.com/fuinorg/event-store-commons)
+- Use @JsonIgnoreProperties(ignoreUnknown = true) on Jackson to be aligned with JSON-B/JAXB which already tolerate unknowns
+- Added a correlation/causation auto-propagation mechanism via [MessageContext](core/src/main/java/org/fuin/ddd4j/core/MessageContext.java) / [WritableMessageContext](core/src/main/java/org/fuin/ddd4j/core/WritableMessageContext.java) / [ThreadLocalMessageContext](core/src/main/java/org/fuin/ddd4j/core/ThreadLocalMessageContext.java)
 
 ## 0.7.0
 - Added [Jackson](jackson) module